refactor: QnAService 리팩토링
